backend/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-multiselectfield==0.1.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.4
whitenoise==6.2.0
